scons: Fix build failures with ASAN enabled.
Currently gem5 will not build on some platforms with the address
sanitizer enabled (`scons --with-asan ...`).

As part of the build process, Scons builds a custom Python
interpreter (`gem5py`) which wraps the same version of Python that will
be embedded in the gem5 binary. This custom interpreter is used later
in the build process.

Currently, when ASAN is enabled, `gem5py` will also be built with ASAN
enabled. Any memory leaks in `gem5py`, or its payload, will then cause
the build to fail when the custom interpreter is used. This patch
updates the SConstruct file to build the gem5py without ASAN even when
ASAN is enabled for the gem5 binary itself.

Change-Id: Ibebfa06a29f4f62990663ae17790c205505880f8